Make filter tag a separate view
